Rename home_page module to homepage
For consistency
