+++ network-api/networkapi/wagtailpages/models.py
@@ -109,7 +109,7 @@ from .pagemodels.buyersguide.article_page import (
     BuyersGuideArticlePageRelatedArticleRelation,
 )
 from .pagemodels.buyersguide.editorial_content_index import BuyersGuideEditorialContentIndexPage
-from .pagemodels.buyersguide.home_page import BuyersGuidePage
+from .pagemodels.buyersguide.homepage import BuyersGuidePage
 from .pagemodels.buyersguide.products import (
     BuyersGuideProductCategory,
     GeneralProductPage,
