+++ network-api/networkapi/wagtailpages/tests/buyersguide/base.py
@@ -5,7 +5,7 @@ from django.test.utils import override_settings
 
 from networkapi.wagtailpages.factory.homepage import WagtailHomepageFactory
 from networkapi.wagtailpages.pagemodels.base import Homepage
-from networkapi.wagtailpages.pagemodels.buyersguide.home_page import BuyersGuidePage
+from networkapi.wagtailpages.pagemodels.buyersguide.homepage import BuyersGuidePage
 from networkapi.wagtailpages.pagemodels.buyersguide.products import (
     ProductPage,
 )
